Split BibleViewModel into focused modules
Refactor: extracted models and major responsibilities from BibleViewModel into separate files — BibleModels.kt, BibleViewModelDetection.kt, BibleViewModelLoading.kt, BibleViewModelSelection.kt and BibleViewModelSmartSearch.kt. Many members changed visibility to internal to enable modular access and testing. Updated imports in MainDesktop, BibleTab and tests to reference moved functions (onEngineScripture, getSelectedVerses, etc.). Purpose: improve separation of concerns (detection, loading, selection, smart search), readability and testability with no intended behavioral changes.
